+++ packages/storybook/stories/constants.ts
@@ -46,8 +46,15 @@ export const compositeExperienceContainerStyle: IStackStyles = {
   }
 };
 
+export const overviewPageImagesStackStyle = { backgroundColor: '#EFF6FC', padding: '2.25rem' };
+
 export const COMPONENT_FOLDER_PREFIX = 'UI Components';
 export const COMPOSITE_FOLDER_PREFIX = 'Composites';
 export const EXAMPLES_FOLDER_PREFIX = 'Examples';
 export const CONCEPTS_FOLDER_PREFIX = 'Concepts';
 export const STATEFUL_CLIENT_PREFIX = 'Stateful Client';
+
+export const MICROSOFT_AZURE_PREVIEWS_URL = 'https://azure.microsoft.com/support/legal/preview-supplemental-terms/';
+
+export const MICROSOFT_AZURE_ACCESS_TOKEN_QUICKSTART =
+  'https://docs.microsoft.com/en-us/azure/communication-services/quickstarts/identity/quick-create-identity';