+++ packages/acs-ui-common/CHANGELOG.md
@@ -1,8 +1,18 @@
 # Change Log - @internal/acs-ui-common
 
-This log was last generated on Tue, 28 Sep 2021 19:19:18 GMT and should not be manually modified.
+This log was last generated on Wed, 17 Nov 2021 22:21:27 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [1.0.0-beta.8](https://github.com/azure/communication-ui-library/tree/@internal/acs-ui-common_v1.0.0-beta.8)
+
+Wed, 17 Nov 2021 22:21:27 GMT 
+[Compare changes](https://github.com/azure/communication-ui-library/compare/@internal/acs-ui-common_v1.0.0-beta.7..@internal/acs-ui-common_v1.0.0-beta.8)
+
+### Changes
+
+- centralized localization utils to acs-ui-common ([PR #974](https://github.com/azure/communication-ui-library/pull/974) by alcail@microsoft.com)
+- Fully implement identifier to/from MRI conversion ([PR #1038](https://github.com/azure/communication-ui-library/pull/1038) by 82062616+prprabhu-ms@users.noreply.github.com)
 
 ## [1.0.0-beta.6](https://github.com/azure/communication-ui-library/tree/@internal/acs-ui-common_v1.0.0-beta.6)